requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pyyaml==6.0.2
tiktoken==0.8.0
pathspec==0.12.1
pydantic==2.10.5
python-multipart==0.0.20
